+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/AbstractConnectOperator.java
@@ -112,11 +112,11 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
     protected final ServiceAccountOperator serviceAccountOperations;
     private final int port;
 
-    private Map<String, Counter> connectorsReconciliationsCounterMap;
-    private Map<String, Counter> connectorsFailedReconciliationsCounterMap;
-    private Map<String, Counter> connectorsSuccessfulReconciliationsCounterMap;
-    private Map<String, AtomicInteger> connectorsResourceCounterMap;
-    private Map<String, Timer> connectorsReconciliationsTimerMap;
+    private Map<String, Counter> connectorsReconciliationsCounterMap = new ConcurrentHashMap<>(1);
+    private Map<String, Counter> connectorsFailedReconciliationsCounterMap = new ConcurrentHashMap<>(1);
+    private Map<String, Counter> connectorsSuccessfulReconciliationsCounterMap = new ConcurrentHashMap<>(1);
+    private Map<String, AtomicInteger> connectorsResourceCounterMap = new ConcurrentHashMap<>(1);
+    private Map<String, Timer> connectorsReconciliationsTimerMap = new ConcurrentHashMap<>(1);
 
     public AbstractConnectOperator(Vertx vertx, PlatformFeaturesAvailability pfa, String kind,
                                    CrdOperator<C, T, L> resourceOperator,
@@ -139,13 +139,6 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
         this.operatorNamespaceLabels = config.getOperatorNamespaceLabels();
         this.pfa = pfa;
         this.port = port;
-
-        connectorsReconciliationsCounterMap = new ConcurrentHashMap<>();
-        connectorsFailedReconciliationsCounterMap = new ConcurrentHashMap<>();
-        connectorsSuccessfulReconciliationsCounterMap = new ConcurrentHashMap<>();
-        connectorsResourceCounterMap = new ConcurrentHashMap<>();
-        connectorsReconciliationsTimerMap = new ConcurrentHashMap<>();
-
     }
 
     @Override
@@ -318,7 +311,7 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
             LOGGER.debugCr(reconciliation, "Setting list of connector plugins in Kafka Connect status");
             connectStatus.setConnectorPlugins(connectorPlugins);
 
-            getConnectorsResourceCounter(namespace).set(desiredConnectors.size());
+            connectorsResourceCounter(namespace).set(desiredConnectors.size());
 
             Set<String> deleteConnectorNames = new HashSet<>(runningConnectorNames);
             deleteConnectorNames.removeAll(desiredConnectors.stream().map(c -> c.getMetadata().getName()).collect(Collectors.toSet()));
@@ -359,15 +352,15 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
                                              boolean useResources, String connectorName, KafkaConnector connector) {
         Promise<Void> reconciliationResult = Promise.promise();
 
-        getConnectorsReconciliationsCounter(reconciliation.namespace()).increment();
+        connectorsReconciliationsCounter(reconciliation.namespace()).increment();
         Timer.Sample connectorsReconciliationsTimerSample = Timer.start(metrics.meterRegistry());
 
         if (connector != null && Annotations.isReconciliationPausedWithAnnotation(connector)) {
 
             return maybeUpdateConnectorStatus(reconciliation, connector, null, null).compose(
                 i -> {
-                    connectorsReconciliationsTimerSample.stop(getConnectorsReconciliationsTimer(reconciliation.namespace()));
-                    getConnectorsSuccessfulReconciliationsCounter(reconciliation.namespace()).increment();
+                    connectorsReconciliationsTimerSample.stop(connectorsReconciliationsTimer(reconciliation.namespace()));
+                    connectorsSuccessfulReconciliationsCounter(reconciliation.namespace()).increment();
                     return Future.succeededFuture();
                 }
             );
@@ -375,13 +368,13 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
 
         reconcileConnector(reconciliation, host, apiClient, useResources, connectorName, connector)
                 .onComplete(result -> {
-                    connectorsReconciliationsTimerSample.stop(getConnectorsReconciliationsTimer(reconciliation.namespace()));
+                    connectorsReconciliationsTimerSample.stop(connectorsReconciliationsTimer(reconciliation.namespace()));
 
                     if (result.succeeded())    {
-                        getConnectorsSuccessfulReconciliationsCounter(reconciliation.namespace()).increment();
+                        connectorsSuccessfulReconciliationsCounter(reconciliation.namespace()).increment();
                         reconciliationResult.complete();
                     } else {
-                        getConnectorsFailedReconciliationsCounter(reconciliation.namespace()).increment();
+                        connectorsFailedReconciliationsCounter(reconciliation.namespace()).increment();
                         reconciliationResult.fail(result.cause());
                     }
                 });
@@ -826,33 +819,30 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
         return secretOperations.reconcile(reconciliation, namespace, KafkaConnectCluster.jmxSecretName(name), null);
     }
 
-    public Counter getConnectorsReconciliationsCounter(String namespace) {
-        return Operator.getCounter(namespace, KafkaConnector.RESOURCE_KIND, metrics, null, connectorsReconciliationsCounterMap,
-                METRICS_PREFIX + "reconciliations",
+    public Counter connectorsReconciliationsCounter(String namespace) {
+        return Operator.getCounter(namespace, KafkaConnector.RESOURCE_KIND, METRICS_PREFIX + "reconciliations", metrics, null, connectorsReconciliationsCounterMap,
                 "Number of reconciliations done by the operator for individual resources");
     }
 
-    public Counter getConnectorsFailedReconciliationsCounter(String namespace) {
-        return Operator.getCounter(namespace, KafkaConnector.RESOURCE_KIND, metrics, null, connectorsFailedReconciliationsCounterMap,
-                METRICS_PREFIX + "reconciliations.failed",
+    public Counter connectorsFailedReconciliationsCounter(String namespace) {
+        return Operator.getCounter(namespace, KafkaConnector.RESOURCE_KIND, METRICS_PREFIX + "reconciliations.failed", metrics, null, connectorsFailedReconciliationsCounterMap,
                 "Number of reconciliations done by the operator for individual resources which failed");
     }
 
-    public Counter getConnectorsSuccessfulReconciliationsCounter(String namespace) {
-        return Operator.getCounter(namespace, KafkaConnector.RESOURCE_KIND, metrics, null, connectorsSuccessfulReconciliationsCounterMap,
-                METRICS_PREFIX + "reconciliations.successful",
+    public Counter connectorsSuccessfulReconciliationsCounter(String namespace) {
+        return Operator.getCounter(namespace, KafkaConnector.RESOURCE_KIND, METRICS_PREFIX + "reconciliations.successful", metrics, null, connectorsSuccessfulReconciliationsCounterMap,
                 "Number of reconciliations done by the operator for individual resources which were successful");
     }
 
-    public AtomicInteger getConnectorsResourceCounter(String namespace) {
-        return Operator.getGauge(namespace, KafkaConnector.RESOURCE_KIND, metrics, null, connectorsResourceCounterMap,
-                METRICS_PREFIX + "resources",
+    public AtomicInteger connectorsResourceCounter(String namespace) {
+        return Operator.getGauge(namespace, KafkaConnector.RESOURCE_KIND, METRICS_PREFIX + "resources",
+                metrics, null, connectorsResourceCounterMap,
                 "Number of custom resources the operator sees");
     }
 
-    public Timer getConnectorsReconciliationsTimer(String namespace) {
-        return Operator.getTimer(namespace, KafkaConnector.RESOURCE_KIND, metrics, null, connectorsReconciliationsTimerMap,
-                METRICS_PREFIX + "reconciliations.duration",
+    public Timer connectorsReconciliationsTimer(String namespace) {
+        return Operator.getTimer(namespace, KafkaConnector.RESOURCE_KIND, METRICS_PREFIX + "reconciliations.duration",
+                metrics, null, connectorsReconciliationsTimerMap,
                 "The time the reconciliation takes to complete");
     }