+++ operator-common/src/main/java/io/strimzi/operator/common/Operator.java
@@ -22,6 +22,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Function;
 
 /**
  * Abstraction of an operator which is driven by resources of a given {@link #kind()}.
@@ -56,10 +57,10 @@ public interface Operator {
      */
     default void reconcileAll(String trigger, String namespace, Handler<AsyncResult<Void>> handler) {
         allResourceNames(namespace).onComplete(ar -> {
-            getPausedResourceCounter(namespace).set(0);
+            pausedResourceCounter(namespace).set(0);
             if (ar.succeeded()) {
                 reconcileThese(trigger, ar.result(), namespace, handler);
-                getPeriodicReconciliationsCounter(namespace).increment();
+                periodicReconciliationsCounter(namespace).increment();
             } else {
                 handler.handle(ar.map((Void) null));
             }
@@ -69,7 +70,7 @@ public interface Operator {
     default void reconcileThese(String trigger, Set<NamespaceAndName> desiredNames, String namespace, Handler<AsyncResult<Void>> handler) {
         if (desiredNames.size() > 0) {
             List<Future> futures = new ArrayList<>();
-            getResourceCounter(namespace).set(desiredNames.size());
+            resourceCounter(namespace).set(desiredNames.size());
 
             for (NamespaceAndName resourceRef : desiredNames) {
                 Reconciliation reconciliation = new Reconciliation(trigger, kind(), resourceRef.getNamespace(), resourceRef.getName());
@@ -77,7 +78,7 @@ public interface Operator {
             }
             CompositeFuture.join(futures).map((Void) null).onComplete(handler);
         } else {
-            getResourceCounter(namespace).set(0);
+            resourceCounter(namespace).set(0);
             handler.handle(Future.succeededFuture());
         }
     }
@@ -99,60 +100,35 @@ public interface Operator {
         return Optional.empty();
     }
 
-    Counter getPeriodicReconciliationsCounter(String namespace);
+    Counter periodicReconciliationsCounter(String namespace);
 
-    AtomicInteger getResourceCounter(String namespace);
+    AtomicInteger resourceCounter(String namespace);
 
-    AtomicInteger getPausedResourceCounter(String namespace);
+    AtomicInteger pausedResourceCounter(String namespace);
 
-    static Counter getCounter(String namespace, String kind, MetricsProvider metrics, Labels selectorLabels, Map<String, Counter> counterMap, String metricName, String metricHelp) {
+    private static <M> M metric(String namespace, String kind, Labels selectorLabels, Map<String, M> metricMap, Function<Tags, M> fn) {
         String selectorValue = selectorLabels != null ? selectorLabels.toSelectorString() : "";
         Tags metricTags = null;
+        String metricKey = namespace + "/" + kind;
         if (namespace.equals("*")) {
             metricTags = Tags.of(Tag.of("kind", kind), Tag.of("selector", selectorValue));
         } else {
             metricTags = Tags.of(Tag.of("kind", kind), Tag.of("namespace", namespace), Tag.of("selector", selectorValue));
         }
+        Tags finalMetricTags = metricTags;
 
-        Counter counter = counterMap.get(namespace + "/" + kind);
-        if (counter == null) {
-            counter = metrics.counter(metricName, metricHelp, metricTags);
-            counterMap.put(namespace + "/" + kind, counter);
-        }
-        return counter;
+        return metricMap.computeIfAbsent(metricKey, x -> fn.apply(finalMetricTags));
     }
 
-    static AtomicInteger getGauge(String namespace, String kind, MetricsProvider metrics, Labels selectorLabels, Map<String, AtomicInteger> gaugeMap, String metricName, String metricHelp) {
-        String selectorValue = selectorLabels != null ? selectorLabels.toSelectorString() : "";
-        Tags metricTags = null;
-        if (namespace.equals("*")) {
-            metricTags = Tags.of(Tag.of("kind", kind), Tag.of("selector", selectorValue));
-        } else {
-            metricTags = Tags.of(Tag.of("kind", kind), Tag.of("namespace", namespace), Tag.of("selector", selectorValue));
-        }
-
-        AtomicInteger gauge = gaugeMap.get(namespace + "/" + kind);
-        if (gauge == null) {
-            gauge = metrics.gauge(metricName, metricHelp, metricTags);
-            gaugeMap.put(namespace + "/" + kind, gauge);
-        }
-        return gauge;
+    static Counter getCounter(String namespace, String kind, String metricName, MetricsProvider metrics, Labels selectorLabels, Map<String, Counter> counterMap, String metricHelp) {
+        return metric(namespace, kind, selectorLabels, counterMap, tags -> metrics.counter(metricName, metricHelp, tags));
     }
 
-    static Timer getTimer(String namespace, String kind, MetricsProvider metrics, Labels selectorLabels, Map<String, Timer> timerMap, String metricName, String metricHelp) {
-        String selectorValue = selectorLabels != null ? selectorLabels.toSelectorString() : "";
-        Tags metricTags = null;
-        if (namespace.equals("*")) {
-            metricTags = Tags.of(Tag.of("kind", kind), Tag.of("selector", selectorValue));
-        } else {
-            metricTags = Tags.of(Tag.of("kind", kind), Tag.of("namespace", namespace), Tag.of("selector", selectorValue));
-        }
+    static AtomicInteger getGauge(String namespace, String kind, String metricName, MetricsProvider metrics, Labels selectorLabels, Map<String, AtomicInteger> gaugeMap, String metricHelp) {
+        return metric(namespace, kind, selectorLabels, gaugeMap, tags -> metrics.gauge(metricName, metricHelp, tags));
+    }
 
-        Timer timer = timerMap.get(namespace + "/" + kind);
-        if (timer == null) {
-            timer = metrics.timer(metricName, metricHelp, metricTags);
-            timerMap.put(namespace + "/" + kind, timer);
-        }
-        return timer;
+    static Timer getTimer(String namespace, String kind, String metricName, MetricsProvider metrics, Labels selectorLabels, Map<String, Timer> timerMap, String metricHelp) {
+        return metric(namespace, kind, selectorLabels, timerMap, tags -> metrics.timer(metricName, metricHelp, tags));
     }
 }